my_app/requirements.txt
streamlit
openai
pandas
python-docx
pypdf
python-pptx
openpyxl
xlsxwriter
xlrd